Move MLD Diagnostics out of MOM_diabatic_aux.F90
Move the calculation of density difference and energy difference
based MLD diagnostics out of MOM_diabatic_aux and into a new module
MOM_diagnose_MLD.F90. Because this new module includes only the
calculation of mixed layer depths that will be primarily used as
diagnostics, it will be in the diagnostics subfolder.
This change will allow the diagnose MLD routines to be used in
the generic tracers.